Move cs_manager tests to own file
